feature: separate favourites table

+++ internal/repository/postgres/recipe_collections.go
@@ -11,7 +11,11 @@ import (
 )
 
 var getRecipeCollectionIdsSubquery = fmt.Sprintf(`
-	SELECT collection_id
+	SELECT COALESCE(
+		jsonb_agg(%[2]v.collection_id)
+		FILTER (WHERE %[2]v.collection_id IS NOT NULL),
+		'[]'
+	)
 	FROM %[2]v
 	LEFT JOIN
 		%[3]v ON %[3]v.collection_id=%[2]v.collection_id
@@ -20,7 +24,7 @@ var getRecipeCollectionIdsSubquery = fmt.Sprintf(`
 	LEFT JOIN
 		%[5]v ON %[5]v.collection_id=%[2]v.collection_id
 	WHERE
-		%[2]v.recipe_id=%[1]v.recipe_id AND %[5]v.user_id=$2 AND (%[3]v.visibility<>'%[6]v' OR %[2]v.contributor_id=$2 AND role='%[7]v')
+		%[2]v.recipe_id=%[1]v.recipe_id AND %[5]v.user_id=$1 AND (%[3]v.visibility<>'%[6]v' OR %[4]v.contributor_id=$1 AND role='%[7]v')
 `, recipesTable, recipesCollectionsTable, collectionsTable, collectionContributorsTable, collectionUsersTable,
 	model.VisibilityPrivate, entity.RoleOwner)
 
@@ -32,17 +36,21 @@ func (r *Repository) SetRecipeCollections(recipeId, userId uuid.UUID, collection
 
 	editableCollections, err := r.getEditableCollections(tx, userId)
 	if err != nil {
-		return errorWithTransactionRollback(tx, fail.GrpcNotFound)
+		return errorWithTransactionRollback(tx, fail.GrpcUnknown)
+	}
+
+	if len(editableCollections) == 0 {
+		return nil
 	}
 
 	clearCollectionsQuery := fmt.Sprintf(`
 		DELETE FROM %s
 		WHERE recipe_id=$1 AND collection_id=ANY($2)
-	`, recipeUsersTable)
+	`, recipesCollectionsTable)
 
 	if _, err := tx.Exec(clearCollectionsQuery, recipeId, editableCollections); err != nil {
 		log.Errorf("unable to clear recipe %s collections for user %s: %s", recipeId, userId, err)
-		return errorWithTransactionRollback(tx, fail.GrpcNotFound)
+		return errorWithTransactionRollback(tx, fail.GrpcUnknown)
 	}
 
 	editableCollectionsMap := make(map[uuid.UUID]bool)
@@ -52,30 +60,36 @@ func (r *Repository) SetRecipeCollections(recipeId, userId uuid.UUID, collection
 
 	setRecipeCollectionsQuery := fmt.Sprintf(`
 		INSERT INTO %[1]v (recipe_id, collection_id)
-		VALUES `, collectionContributorsTable)
+		VALUES `, recipesCollectionsTable)
 
 	var args []interface{}
 	for _, collectionId := range collections {
 		if _, ok := editableCollectionsMap[collectionId]; !ok {
 			continue
 		}
-		setRecipeCollectionsQuery += "(?, ?),"
+		currentIndex := len(args) + 1
+		setRecipeCollectionsQuery += fmt.Sprintf("($%[1]v, $%[2]v),", currentIndex, currentIndex+1)
 		args = append(args, recipeId, collectionId)
 	}
-	setRecipeCollectionsQuery = setRecipeCollectionsQuery[0 : len(setRecipeCollectionsQuery)-1]
-	setRecipeCollectionsQuery += fmt.Sprint(" ON CONFLICT DO NOTHING")
 
-	if _, err := tx.Exec(setRecipeCollectionsQuery, args); err != nil {
-		log.Errorf("unable to set recipe %s collections for user %s: %s", recipeId, userId, err)
-		return errorWithTransactionRollback(tx, fail.GrpcNotFound)
+	if len(args) == 0 {
+		return commitTransaction(tx)
 	}
 
-	return nil
+	setRecipeCollectionsQuery = setRecipeCollectionsQuery[0 : len(setRecipeCollectionsQuery)-1]
+	setRecipeCollectionsQuery += fmt.Sprint(" ON CONFLICT (recipe_id, collection_id) DO NOTHING")
+
+	if _, err = tx.Exec(setRecipeCollectionsQuery, args...); err != nil {
+		log.Errorf("unable to set recipe %s collections for user %s: %s", recipeId, userId, err)
+		return errorWithTransactionRollback(tx, fail.GrpcUnknown)
+	}
+
+	return commitTransaction(tx)
 }
 
 func (r *Repository) getEditableCollections(tx *sql.Tx, userId uuid.UUID) ([]uuid.UUID, error) {
 	var getEditableCollectionsQuery = fmt.Sprintf(`
-		SELECT collection_id
+		SELECT %[1]v.collection_id
 		FROM %[1]v
 		LEFT JOIN
 			%[2]v ON %[2]v.collection_id=%[1]v.collection_id
